pos-edi/services/outbox-worker/requirements.txt
pymysql
SQLAlchemy==2.0.32
kafka-python==2.0.2
python-dotenv==1.0.1
